pyproject.toml
[project]
name = "lexcql"
version = "1.0.0"
description = "LexCQL Query Grammar and Parser"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Erik Körner", email = "koerner@saw-leipzig.de" }
]
keywords = [
    "LexCQL",
    "FCS",
    "CQL",
    "Query Parser"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: General",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
]

[project.urls]
Source = "https://github.com/Querela/lexcql-python.git"
Issues = "https://github.com/Querela/lexcql-python/issues"

[project.optional-dependencies]
antlr = [
    "antlr4-tools>=0.2.2",
]
test = [
    "pytest>=9.0.2",
    "pytest-clarity>=1.0.1",
    "pytest-cov>=7.0.0",
    "pytest-randomly>=4.0.1",
]
style = [
    "black>=26.1.0",
    "flake8>=7.3.0",
    "isort>=8.0.0",
]
build = [
    "twine>=6.2.0",
]

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
source-exclude = [".antlr"]

[tool.black]
line-length = 120
extend-exclude = '''
(
  (Lex(Lexer|Parser(Listener)?)).py
)
'''

[tool.isort]
force_single_line = true
line_length = 120
known_first_party = "lexcql"
default_section = "THIRDPARTY"
forced_separate = "test_lexcql"
skip = [
    ".venv",
    "dist",
    "src/lexcql/LexParser.py",
    "src/lexcql/LexLexer.py",
    "src/lexcql/LexParserListener.py",
]
skip_gitignore = true
